+++ components/steps/w-step.vue
@@ -1,7 +1,15 @@
 <script setup>
 import { computed, inject } from 'vue';
+import { i18n } from '@lingui/core';
+import { activateI18n } from '../util/i18n';
 import { step as ccStep } from '@warp-ds/css/component-classes';
 import { IconCheck16 } from "@warp-ds/icons/vue";
+import { messages as enMessages} from './locales/en/messages.mjs';
+import { messages as nbMessages} from './locales/nb/messages.mjs';
+import { messages as fiMessages} from './locales/fi/messages.mjs';
+
+
+activateI18n(enMessages, nbMessages, fiMessages);
 
 const vertical = inject('steps-vertical', true);
 const left = inject('steps-left', true);
@@ -10,6 +18,35 @@ const props = defineProps({
   active: Boolean,
   complete: Boolean,
 });
+
+const availableAriaLabels = {
+  complete: i18n._(
+    /*i18n*/ {
+      id: "steps.aria.completed",
+      message: "Step indicator completed circle",
+      comment: "Completed circle",
+    }
+  ),
+  active: i18n._(
+    /*i18n*/ {
+      id: "steps.aria.active",
+      message: "Step indicator active circle",
+      comment: "Active circle",
+    }
+  ),
+  default: i18n._(
+    /*i18n*/ {
+      id: "steps.aria.emptyCircle",
+      message: "Empty circle",
+      comment: "Empty circle",
+    }
+  )
+}
+
+const getAriaLabel = (props) => {
+  const ariaLabel = Object.keys(availableAriaLabels).find(a => props[a]);
+  return ariaLabel ? availableAriaLabels[ariaLabel] : availableAriaLabels.default;
+}
 
 const stepClasses = computed(() => [
   ccStep.step,
@@ -66,16 +103,16 @@ const contentClasses = computed(() => [
 </script>
 
 <template>
-  <div :class="stepClasses">
+  <li :class="stepClasses">
     <div v-if="!vertical" :class="horizontalClasses" />
-    <div :aria-current="active ? 'step' : undefined" :class="stepDotClasses">
+    <div role="img" :aria-label="getAriaLabel(props)" :aria-current="active ? 'step' : undefined" :class="stepDotClasses">
       <icon-check-16 v-if="complete" />
     </div>
     <div :class="stepLineClasses" />
     <div :class="contentClasses">
       <slot />
     </div>
-  </div>
+  </li>
 </template>
 
 <script>